リファクタリング取込
