+++ frontend/features/notes/components/noteInput.tsx
