+++ frontend/package.json
@@ -18,12 +18,10 @@
     "@emotion/styled": "^11.13.0",
     "@hookform/resolvers": "^3.6.0",
     "@supabase/supabase-js": "^2.45.4",
-    "autoprefixer": "^10.4.19",
     "axios": "^1.7.7",
     "dotenv": "^16.4.5",
     "framer-motion": "^11.11.8",
     "next": "^14.2.21",
-    "postcss": "^8.4.40",
     "react": "^18.3.1",
     "react-datepicker": "^7.3.0",
     "react-dom": "^18.3.1",
@@ -32,7 +30,6 @@
     "react-router-dom": "^6.23.1",
     "react-textarea-autosize": "^8.5.3",
     "swr": "^2.2.5",
-    "tailwindcss": "^3.4.7",
     "use-debounce": "^10.0.1",
     "zod": "^3.23.8"
   },
