+++ frontend/pages/_app.tsx
@@ -1,8 +1,7 @@
-import "../styles/globals.css";
 import React from "react";
 import type { AppProps } from "next/app";
 import { ChakraProvider } from "@chakra-ui/react";
-import { AuthProvider } from "../context/AuthContext";
+import { AuthProvider } from "../../frontend/features/auth/AuthContext";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
