+++ frontend/pages/forgot-password.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ForgotPassword from '../components/pages/ForgotPassword';
+import ForgotPassword from '../features/auth/components/ForgotPasswordPage';
 
 const ForgotPasswordPage: React.FC = () => {
   return <ForgotPassword />;
