+++ frontend/pages/top.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect } from 'react';
 import { useRouter } from 'next/router';
-import { useAuth } from '../context/AuthContext';
+import { useAuth } from '../../frontend/features/auth/AuthContext';
 import { Spinner, Center } from '@chakra-ui/react';
-import Top from '../components/pages/top';
+import Top from '../../frontend/features/top/components/TopPage';
 
 const TopPage: React.FC = () => {
 //   const { user } = useAuth();
