--- frontend/postcss.config.mjs
@@ -1,9 +1,0 @@
-/** @type {import('postcss-load-config').Config} */
-const config = {
-  plugins: {
-    tailwindcss: {},
-    autoprefixer: {}, 
-  },
-};
-
-export default config;
