+++ frontend/theme.js
@@ -1,4 +1,3 @@
-// theme.js
 import { extendTheme } from "@chakra-ui/react";
 
 const theme = extendTheme({
